pyproject.toml
[project]
name = "openephys-to-dh"
version = "0.1.0"
description = "Python tools to convert and preprocess OpenEphys data to the DAQ-HDF5 (dh5) format"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dataclasses-json>=0.6.7",
    "dh-format",
    "open-ephys-python-tools",
    "scipy>=1.15.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true

[tool.uv.sources]
dh-format = { git = "https://github.com/cog-neurophys-lab/dh5io.git" }
open-ephys-python-tools = { git = "https://github.com/joschaschmiedt/open-ephys-python-tools.git", branch = "add-tests" }


[dependency-groups]
dev = ["pytest>=8.3.5"]


[project.scripts]
oe_to_dh = "cli.oe_to_dh:main"
dhtree = "cli.dhtree:main"
